% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mcvis.R
\name{ggplot_mcvis}
\alias{ggplot_mcvis}
\title{ggplot visualisation for mcvis method}
\usage{
ggplot_mcvis(mcvis_result, eig.max = ncol(mcvis_result$MC),
  var.max = ncol(mcvis_result$MC))
}
\arguments{
\item{mcvis_result}{Output of the mcvis function}

\item{eig.max}{The maximum number of eigenvalues to be displayed on the plot.}

\item{var.max}{The maximum number of variables (i.e. columns) to be displayed on the plot.}
}
\value{
A ggplot
}
\description{
The main output of the `mcvis` function is the MC-index matrix,
which has the tau's as rows and the original variables as columns.
The ggplot_mcvis function first orders the matrix columns by
the magnitude of the MC-index for the tau1, which is the inverse of the smallest eigenvalue
Under this ordering, the main cause of collinearity is the first entry in the ordered matrix.
Plotting parameters (e.g. thickness and categories) are then chosen to highlight.
}
\examples{
set.seed(1)
p = 10
n = 100
X = matrix(rnorm(n*p), ncol = p)
X[,1] = X[,2] + X[,3] + rnorm(n, 0, 0.1)
mcvis_result = mcvis(X)
ggplot_mcvis(mcvis_result)
}
\author{
Chen Lin, Kevin Wang, Samuel Mueller
}
