% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_mcvis.R
\name{igraph_mcvis}
\alias{igraph_mcvis}
\title{Multi-collinearity Visualization}
\usage{
igraph_mcvis(mcvis_result, eig.max = ncol(g), vol.max = ncol(g))
}
\arguments{
\item{mcvis_result}{Output of the mcvis function}

\item{eig.max}{The maximum number of eigenvalues to be displayed on the plot.}

\item{vol.max}{The maximum number of variables to be displayed on the plot.}
}
\description{
Multi-collinearity Visualization
}
\examples{
library(mplot)
data("artificialeg")
p=dim(artificialeg)[2]-1
X=artificialeg[,1:p]
mcvis_result = mcvis2(X)
igraph_mcvis(mcvis_result)
}
\author{
Chen Lin
}
