% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcvis.R
\name{mcvis}
\alias{mcvis}
\title{Multi-collinearity Visualization}
\usage{
mcvis(X, method = "bootstrap", steps = 1000L)
}
\arguments{
\item{X}{A matrix of regressors (without intercept terms).}

\item{method}{The resampling method for the data. Current supports "bootstrap" or "cv"(cross-validation).}

\item{steps}{Number of sampling runs we perform. Default to 1000.}
}
\description{
Multi-collinearity Visualization
}
\examples{
library(mplot)
data("artificialeg")
X=artificialeg[,1:9]
mcvis(X)
}
\author{
Chen Lin, Kevin Wang
}
