% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcvis2.R
\name{mcvis2}
\alias{mcvis2}
\title{Multi-collinearity Visualization}
\usage{
mcvis2(X, eig.max = dim(X)[2], vol.max = dim(X)[2],
  method = "bootstrap", steps = 1000L)
}
\arguments{
\item{X}{A matrix of regressors.}

\item{eig.max}{The maximum number of eigenvalues to be displayed on the plot.}

\item{vol.max}{The maximum number of variables to be displayed on the plot.}

\item{method}{The resampling method for the data. Current supports "bootstrap" or "cv"(cross-validation).}

\item{steps}{Number of sampling runs we perform. Default to 1000.}

\item{tau}{A parameter for plotting line thickness}
}
\description{
Multi-collinearity Visualization
}
\examples{
library(mplot)
data("artificialeg")
p=dim(artificialeg)[2]-1
X=artificialeg[,1:p]
mcvis2(X)
}
\author{
Chen Lin, Kevin Wang
}
