% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMcvis.R
\name{oldMcvis}
\alias{oldMcvis}
\title{Multi-collinearity Visualization}
\usage{
oldMcvis(X, tau = 1.5, col.names, firstcol = FALSE,
  eig.max = dim(X)[2] - firstcol, vol.max = dim(X)[2] - firstcol,
  method = "bootstrap")
}
\arguments{
\item{X}{a matrix of regressors}

\item{col.names}{the name of variables}

\item{firstcol}{logical; if the first column of X is constant of ones.}

\item{eig.max}{the maximum number of eigenvalues to be displayed on the plot}

\item{vol.max}{the maximum number of variables to be displayed on the plot}

\item{method}{"bootstrap" or "cv"(cross-validation); the method used to resample the data.}
}
\description{
Multi-collinearity Visualization
}
\examples{
library(mplot)
data("artificialeg")
p=dim(artificialeg)[2]-1
X=artificialeg[,1:p]
steps = 1000L
oldMcvis(X)
}
\author{
Chen Lin
}
