% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mcvis.R
\name{shiny_mcvis}
\alias{shiny_mcvis}
\title{Shiny app for mcvis exploration}
\usage{
shiny_mcvis()
}
\arguments{
\item{mcvis_result}{Output of the mcvis function}

\item{eig.max}{The maximum number of eigenvalues to be displayed on the plot.}

\item{vol.max}{The maximum number of variables to be displayed on the plot.}
}
\description{
Shiny app for mcvis exploration
}
\examples{
library(mplot)
data("artificialeg")
p=dim(artificialeg)[2]-1
X=artificialeg[,1:p]
mcvis_result = mcvis2(X)
shiny_mcvis()
}
\author{
Chen Lin, Kevin Wang
}
