% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{Visualization of collinearty}
\usage{
visualization(Y, X, tau = 1.5, steps = 1000, p = dim(X)[2] - 1,
  col.names = colnames(X)[2:(p + 1)], eig.max = p, vol.max = p,
  method = "bootstrap")
}
\arguments{
\item{Y}{respose numeric vector}

\item{X}{design matrix}
}
\description{
Visualization of collinearty
}
\examples{
library(mplot)
data("artificialeg")
n=dim(artificialeg)[1]
p=dim(artificialeg)[2]-1
n1=as.matrix(rep(1,n))
X=as.matrix(cbind(n1,artificialeg[,1:p]))
Y=artificialeg[,p+1]
visualization(Y,X)
}
\author{
Chen Lin
}
