% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mcvis.R
\name{ggplot_mcvis}
\alias{ggplot_mcvis}
\title{ggplot visualisation for mcvis method}
\usage{
ggplot_mcvis(mcvis_result, eig.max = ncol(mcvis_result$MC),
  var.max = ncol(mcvis_result$MC))
}
\arguments{
\item{mcvis_result}{Output of the mcvis function}

\item{eig.max}{The maximum number of eigenvalues to be displayed on the plot.}

\item{var.max}{The maximum number of variables (i.e. columns) to be displayed on the plot.}
}
\value{
A ggplot
}
\description{
The ggplot_mcvis function first orders the MC-index matrix columns by
the magnitude of the MC-index for the tau1, which is the inverse of the smallest eigenvalue
Under this ordering,
the first entry of the matrix is classified as a "category 5" thickness in plotting
Subsequently, other lines are sequentially put in the categories of
\itemize{
 \item category 5: anything equal or above the first entry of the matrix
 \item category 4: anything above 90 \% of the first entry of the matrix
 \item category 3: anything above 70 \% of the first entry of the matrix
 \item category 2: anything above 50 \% of the first entry of the matrix
 \item category 1: anything above 1 \% of the first entry of the matrix
}
}
\examples{
set.seed(1)
p = 10
n = 100
X = matrix(rnorm(n*p), ncol = p)
X[,1] = X[,2] + rnorm(n, 0, 0.1)
mcvis_result = mcvis(X)
ggplot_mcvis(mcvis_result)
}
\author{
Chen Lin, Kevin Wang, Samuel Mueller
}
