% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mcvis.R
\name{ggplot_mcvis}
\alias{ggplot_mcvis}
\title{ggplot visualisation for mcvis method}
\usage{
ggplot_mcvis(mcvis_result, eig.max = ncol(mcvis_result$MC),
  vol.max = ncol(mcvis_result$MC))
}
\arguments{
\item{mcvis_result}{Output of the mcvis function}

\item{eig.max}{The maximum number of eigenvalues to be displayed on the plot.}

\item{vol.max}{The maximum number of variables to be displayed on the plot.}
}
\description{
ggplot visualisation for mcvis method
}
\examples{
library(mplot)
data("artificialeg")
X=artificialeg[,1:9]
mcvis_result = mcvis(X)
ggplot_mcvis(mcvis_result)
}
\author{
Chen Lin, Kevin Wang
}
