% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_mcvis.R
\name{igraph_mcvis}
\alias{igraph_mcvis}
\title{Multi-collinearity Visualization}
\usage{
igraph_mcvis(mcvis_result, tau = 2/3, eig.max = ncol(mcvis_result$g),
  vol.max = ncol(mcvis_result$g))
}
\arguments{
\item{mcvis_result}{Output of the mcvis function}

\item{tau}{A parameter determining below which thickness the plotting lines hide.}

\item{eig.max}{The maximum number of eigenvalues to be displayed on the plot.}

\item{vol.max}{The maximum number of variables to be displayed on the plot.}
}
\description{
Multi-collinearity Visualization
}
\examples{
library(mplot)
data("artificialeg")
X=artificialeg[,1:9]
mcvis_result = mcvis(X)
igraph_mcvis(mcvis_result)
}
\author{
Chen Lin
}
