% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcvis.R
\name{mcvis}
\alias{mcvis}
\title{Multi-collinearity Visualization}
\usage{
mcvis(X, tau = 1.5, col.names, col.one = FALSE, eig.max = dim(X)[2] -
  col.one, vol.max = dim(X)[2] - col.one, method = "bootstrap")
}
\arguments{
\item{X}{design matrix}

\item{Y}{respose numeric vector}
}
\description{
Multi-collinearity Visualization
}
\examples{
library(mplot)
data("artificialeg")
p=dim(artificialeg)[2]-1
X=artificialeg[,1:p]
mcvis(X)
}
\author{
Chen Lin
}
