% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcvis.R
\name{mcvis}
\alias{mcvis}
\title{Multi-collinearity Visualization}
\usage{
mcvis(X, tau = 1.5, steps = 1000, col.names, eig.max, vol.max,
  method = "bootstrap", col.one = FALSE)
}
\arguments{
\item{X}{design matrix}

\item{Y}{respose numeric vector}
}
\description{
Multi-collinearity Visualization
}
\examples{
library(mplot)
data("artificialeg")
p=dim(artificialeg)[2]-1
X=artificialeg[,1:p]
mcvis(X)
}
\author{
Chen Lin
}
