% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcvis.R
\name{mcvis}
\alias{mcvis}
\title{Multi-collinearity Visualization}
\usage{
mcvis(X, tau = 1.5, firstcol = FALSE, eig.max = dim(X)[2] - firstcol,
  vol.max = dim(X)[2] - firstcol, method = "bootstrap",
  steps = 1000L)
}
\arguments{
\item{X}{A matrix of regressors.}

\item{tau}{A parameter for plotting line thickness}

\item{firstcol}{Logical; if the first column of X is constant of ones.}

\item{eig.max}{The maximum number of eigenvalues to be displayed on the plot.}

\item{vol.max}{The maximum number of variables to be displayed on the plot.}

\item{method}{The resampling method for the data. Current supports "bootstrap" or "cv"(cross-validation).}

\item{steps}{Number of sampling runs we perform. Default to 1000.}

\item{col.names}{The name of variables.}
}
\description{
Multi-collinearity Visualization
}
\examples{
library(mplot)
data("artificialeg")
p=dim(artificialeg)[2]-1
X=artificialeg[,1:p]
mcvis(X)
}
\author{
Chen Lin
}
