% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcvis.R
\name{mcvis}
\alias{mcvis}
\title{Multi-collinearity Visualization}
\usage{
mcvis(X, method = "bootstrap", steps = 1000L, k = 10L)
}
\arguments{
\item{X}{A matrix of regressors (without intercept terms).}

\item{method}{The resampling method for the data.
Currently supports "bootstrap" or "cv" (cross-validation).}

\item{steps}{Number of resampling runs we perform. Default is set to 1000.}

\item{k}{Number of partitions in averaging theMC index. Default is set to 10.}
}
\value{
A list of outputs:
\item{X: The original matrix of regressors (for plotting purposes)}
\item{t_square: The t^2 statistics for the regression between the VIFs and the tau's.}
\item{MC: The MC indices}
\item{col.names: Column names (for plotting purposes)}
}
\description{
Multi-collinearity Visualization
}
\examples{
library(mplot)
data("artificialeg")
X=artificialeg[,1:9]
mcvis(X)
}
\author{
Chen Lin, Kevin Wang, Samuel Mueller
}
