% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mcvis.R
\name{shiny_mcvis}
\alias{shiny_mcvis}
\title{Shiny app for mcvis exploration}
\usage{
shiny_mcvis(mcvis_result)
}
\arguments{
\item{mcvis_result}{Output of the mcvis function.}
}
\description{
Shiny app for mcvis exploration
}
\examples{
**## Not run:**
library(mplot)
data("artificialeg")
p = dim(artificialeg)[2]-1
X = artificialeg[,1:p]
mcvis_result = mcvis(X)
shiny_mcvis(mcvis_result)
## End(**Not run**)
}
\author{
Chen Lin, Kevin Wang
}
