% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mcvis.R
\name{shiny_mcvis}
\alias{shiny_mcvis}
\title{Shiny app for mcvis exploration}
\usage{
shiny_mcvis(mcvis_result, X)
}
\arguments{
\item{mcvis_result}{Output of the mcvis function}

\item{X}{The original X matrix}
}
\description{
Shiny app for mcvis exploration
}
\examples{
if(interactive()){
set.seed(1)
p = 10
n = 100
X = matrix(rnorm(n*p), ncol = p)
mcvis_result = mcvis(X)
shiny_mcvis(mcvis_result = mcvis_result, X = X)
}
}
\author{
Chen Lin, Kevin Wang, Samuel Mueller
}
